Refactor user checkin polling into separate module
